Extract Common Logic for WireMock Test Setup to Share Across Test Cases
